+++ src/tools/graph_tools.h
@@ -4,7 +4,7 @@
 #include <vector>
     using std::vector;
 
-int count_num_paths(vector<vector<float> > admat, int node1, int node2);
+long unsigned int count_num_paths(vector<vector<float> > admat, int node1, int node2);
